updated files and sections
