+++ package.json
@@ -13,6 +13,7 @@
     "react-dom": "^16.4.0",
     "react-router-dom": "^4.3.1",
     "react-scripts": "1.1.4",
+    "react-scroll-parallax": "^1.3.5",
     "react-simple-card": "^1.0.11",
     "reactstrap": "^6.1.0"
   },
