+++ src/containers/Dashboard/Dashboard.js
@@ -5,7 +5,6 @@ import NavBar from './NavBar/NavBar';
 class Dashboard extends Component {
   render() {
 
-    const userName = (JSON.parse(sessionStorage.getItem("ActiveUser")))
     return (
 
           <NavBar />
