--- src/containers/Dashboard/Friends/Friends.css
@@ -1,60 +1,0 @@
-.list-group {
-    margin: auto;
-    width: 100%;
-}
-
-.profileImage {
-    width: 50% !important;
-    height: 80% !important;
-    opacity: .6;
-    margin: 10px;
-}
-
-.vertical-align {
-    display: flex;
-    align-items: center;
-    height: 87px;
-}
-
-.friendItem {
-    font-size: 1.3em !important;
-}
-
-.userItem {
-font-size: 1.3em !important;
-}
-
-.removeFriendButton {
-    width: 60px;
-    background-color: #E2B6E4;
-    color: white;
-    font-size: 1.4em !important;
-}
-
-.addFriendButton {
-    width: 60px;
-    background-color: #8BA66A !important;
-    color: white;
-    font-size: 1.4em !important;
-}
-
-.friendsView .list-group-item {
-    margin: 5px;
-    background-color: #8BA66A !important;
-    color: white;
-}
-
-.list-group-item {
-    margin: 5px;
-    padding: 0 !important;
-    color: #9B9B9B;
-}
-
-.Friends-title {
-    color: #9B9B9B;
-}
-
-.Users-title {
-    color: #9B9B9B;
-}
-
