+++ src/containers/Dashboard/NavBar/NavBar.css
@@ -7,25 +7,40 @@
 
   .logo {
     display: flex;
-    height: 30px !important;
+    height: 50px !important;
     margin-right: 15px;
     margin-top: 10px;
   }
 
   li {
-    font-size: 1.5em;
-    color: #9B9B9B;
+    font-size: 1.2sem;
+    /* color: #9B9B9B; */
   }
 
   a {
-      color: #9B9B9B !important;
+	  color: white !important;
+	  text-shadow: none !important;
   }
 
   .nav-tabs > li.active > a{
       color: #5fd0b6 !important;
   }
 
-  .navbarHeight {
+  /* .navbarHeight {
       padding-top: 20px;
       padding-bottom: 20px;
-  }
+  } */
+
+  .navContainer {
+    margin: auto;
+	background-color: rgba(66, 153, 159, .8);
+	border: solid 2px #6bc1f7;
+    background-size:20px 20px;
+    background-position:0 0, 10px 10px;
+    height: 80px;
+}
+
+.navbarStyle {
+	display: flex;
+	text-align: center !important;
+}
