+++ src/containers/Dashboard/Title/Title.css
@@ -1,0 +1,9 @@
+h1 {
+	color: white;
+	font-size: 2em;
+}
+
+.titleLogo {
+    height: 200px;
+    margin: 0;
+  }
