+++ src/containers/homePage/Home.css
@@ -2,71 +2,62 @@
 @import url('https://fonts.googleapis.com/css?family=Amatic+SC');
 
 h1 {
-    color: #9B9B9B;
+    color: white;
     font-size: 30px;
 }
 
 body {
-    font-family: 'Amatic SC', cursive;
+    font-family: 'Bangla Sangam MN';
 }
 
-.logoContainer {
-    margin: auto;
-    background-color: rgba(171, 248, 231, .7);
-    background-image: radial-gradient(closest-side, transparent 99%, rgba(0,0,0,.10) 99%),
-    radial-gradient(closest-side, transparent 99%, rgba(0,0,0,.1) 99%);
-    background-size:20px 20px;
-    background-position:0 0, 10px 10px;
-    height: 80px;
-}
 
 .mainLogo {
     height: 40px;
     margin-top: 20px;
 }
 
-.mainSection {
-    background-color:rgba(171, 248, 231, .7);
-    height: 330px;
+.titleSection {
+	background-image: url('../../components/IMAGES/GrayFlowersTop.png');
+	background-size: cover;
+	background-repeat: no-repeat;
+	height: 500px;
+	max-width: 100%;
 }
 
-.mainSection>h1 {
+.titleSection>h1 {
     float: left;
     margin: auto;
 }
 
 .secondSection {
-    display: flex;
-    height: 400px;
-    background-image: url('../../components/IMAGES/sharingtoys.jpg');
-    -webkit-filter: grayscale(100%);
-    z-index: -1;
-    opacity: .5;
-    background-size: 100% auto;
-}
-
-.secondSectionText>h3 {
-    line-height: 2em;
-    color: #E2B6E4;
-    background-color: gray;
-    font-size: 29px;
-    font-weight: 500;
-    margin: 0 0;
+	height: 500px;
+	background-color: rgba(66, 153, 159, .8);
+	border: solid 2px #6bc1f7;
+    background-size:20px 20px;
+    background-position:0 0, 10px 10px;
 }
 
 .thirdSection {
-    display: flex;
-    height: 320px;
-    background-color: rgba(182, 250, 235, 0.7);
+    height: 600px;
+    background-image: url('../../components/IMAGES/GrayFlowersBottom.png');
+	background-size: cover;
+	background-repeat: no-repeat;
 }
 
-.thirdSection>h3 {
-    color: #9B9B9B;
-    padding: 28px;
-    line-height: 1.5em;
-    width: 415px;
-    margin-top: 42px;
-    text-align: center;
+.fourthSection {
+	height: 500px;
+	background-color: rgba(66, 153, 159, .8);
+	border: solid 2px #6bc1f7;
+    background-size:20px 20px;
+    background-position:0 0, 10px 10px;
+}
+
+.fifthSection {
+    height: 600px;
+    background-image: url('../../components/IMAGES/brownfloral.png');
+	background-size: cover;
+	background-repeat: no-repeat;
+	opacity: .7;
 }
 
 .promo {
@@ -91,25 +82,13 @@ body {
     margin: 0 10px;
 }
 
-/* Screen shot of the borrow app */
-.borrowView {
-    height: 220px;
-    border: solid gray 2px;
-    margin: 50px;
-}
-
-.borrowViewText {
-    font-size: 28px;
-}
 
 .footer {
     height: 100px;
-    background-color:  rgba(182, 250, 235, 0.7);
-    background-image: radial-gradient(closest-side, transparent 99%, rgba(0,0,0,.10) 99%),
-    radial-gradient(closest-side, transparent 99%, rgba(0,0,0,.1) 99%);
+	background-color: rgba(66, 153, 159, .8);
+	border: solid 2px #6bc1f7;
     background-size:20px 20px;
     background-position:0 0, 10px 10px;
-    color: #9B9B9B;
 }
 
 .footer>p {
@@ -120,7 +99,7 @@ body {
 .socialicon {
     height: 40px;
     width:  40px;
-    margin: 5px;
+    margin: 0;
 }
