--- src/containers/homePage/register_login/LoginRegisterUser.css
@@ -1,122 +1,0 @@
-/* Large arrow register form container */
-.loginView {
-    float: right;
-    margin-right: 10%;
-    margin-top: 7%;
-    display: inline-grid;
-    background-image: url('../../../components/IMAGES/arrow.png');
-    background-repeat: no-repeat;
-    width: 275px;
-    height:235px;
-}
-
-.loginView:hover{
-    display: inline-grid;
-    background-image: url('../../../components/IMAGES/hoveredarrow.png');
-    background-repeat: no-repeat;
-    background-position-X: -2px;
-    width: 275px;
-    height:235px;
-}
-
-.loginView:active{
-    display: inline-grid;
-    background-image: url('../../../components/IMAGES/pressedarrow.png');
-    background-repeat: no-repeat;
-    background-position-X: 3px;
-    width: 275px;
-    height:235px;
-    /* animation-duration: 1s; */
-}
-
-.loginView:visited{
-    display: inline-grid;
-    background-image: url('../../../components/IMAGES/pressedarrow.png');
-    background-repeat: no-repeat;
-    width: 275px;
-    height:235px;
-    /* animation-duration: 1s; */
-
-}
-
-/* Small arrow login form container */
-.loginFields{
-    background-image: url('../../../components/IMAGES/arrow2.png');
-    background-repeat: no-repeat;
-    width: 255px;
-    padding-bottom: 25px;
-    margin-top: -41px;
-    position: relative;
-    margin-left: 20px;
-}
-
-.loginFields:hover{
-    background-image: url('../../../components/IMAGES/hoveredarrow2.png');
-    background-repeat: no-repeat;
-    background-position-y: -2px;
-}
-
-
-.loginFieldsWrapper {
-    padding-top: 28px;
-    display: -ms-flexbox;
-    display: flex;
-    -ms-flex-direction: column;
-    flex-direction: column;
-    max-width: 160px;
-    margin-left: 10px;
-}
-
-.loginFieldsWrapper input, .registerFieldsWrapper input {
-    margin: 2px 0;
-    color:#E2B6E4;
-    font-size: 16px;
-}
-
-/* Input fields for login */
-.loginForm{
-    width: 80%;
-    height: 80%;
-    margin-left: 9.5%;
-    margin-top: 8%;
-}
-
-/* Input fields for register */
-.registerFieldsWrapper {
-    display: flex;
-    flex-direction: column;
-    max-width: 160px;
-    margin-left: 28px;
-}
-
-/* Sign up button */
-.signUp{
-    background-color: transparent !important;
-    margin-top: 13px;
-    margin-left: 3px;
-    border: none;
-    box-shadow: none;
-    color: white;
-    font-weight: 800;
-    font-size: 15px;
-    width: inherit;
-    outline: none;
-}
-
-/* Sign in button */
-.signIn{
-    background-color: transparent !important;
-    border: none;
-    box-shadow: none;
-    color: white;
-    font-weight: 800;
-    font-size: 15px;
-    width: 40px;
-    display: block;
-    margin-left: 191px;
-    position: absolute;
-    line-height: 1em;
-    outline: none;
-    top: 47px;
-    right: 22px;
-}
